requirements.txt

# Additional requirements go here
factory-boy==2.12.0
faker==4.1.1
